+++ client/src/components/index.js
@@ -13,4 +13,4 @@ export {default as Adventure} from './Adventure'
 export {default as Activity} from './Activity'
 export {default as Pod} from './Pod'
 export {default as Search} from './Search'
-export {default as PollMap} from './PollMap'
+export {default as Map} from './Map'
